+++ src/neuroconv/datainterfaces/behavior/video/videodatainterface.py
@@ -28,7 +28,13 @@ class VideoInterface(BaseDataInterface):
     # Other suffixes, while they can be opened by OpenCV, are not supported by DANDI so should probably not list here
     info = "Interface for handling standard video file formats."
 
-    def __init__(self, file_paths: list, verbose: bool = False):  # TODO - debug why List[FilePathType] fails
+    def __init__(
+        self,
+        file_paths: list,
+        verbose: bool = False,
+        *,
+        metadata_key_name: str = "Videos",
+    ):
         """
         Create the interface for writing videos as ImageSeries.
 
@@ -37,12 +43,37 @@ class VideoInterface(BaseDataInterface):
         file_paths : list of FilePathTypes
             Many video storage formats segment a sequence of videos over the course of the experiment.
             Pass the file paths for this videos as a list in sorted, consecutive order.
+        metadata_key_name : str, optional
+            The key used to identify this video data within the overall experiment metadata.
+            Defaults to "Videos".
+
+            This key is essential when multiple video streams are present in a single experiment.
+            The associated metadata should be a list of dictionaries, with each dictionary
+            containing metadata for a specific video segment:
+
+            ```
+            metadata["Behavior"][metadata_key_name] = [
+                {video1_metadata},
+                {video2_metadata},
+                ...
+            ]
+            ```
+
+            If other video interfaces exist, they would follow a similar structure:
+
+            ```
+            metadata["Behavior"]["other_video_key_name"] = [
+                {other_video1_metadata},
+                {other_video2_metadata},
+                ...
+            ]
         """
         get_package(package_name="cv2", installation_instructions="pip install opencv-python-headless")
         self.verbose = verbose
         self._number_of_files = len(file_paths)
         self._timestamps = None
         self._segment_starting_times = None
+        self.metadata_key_name = metadata_key_name
         super().__init__(file_paths=file_paths)
 
     def get_metadata_schema(self):
@@ -53,26 +84,22 @@ class VideoInterface(BaseDataInterface):
         for key in exclude:
             image_series_metadata_schema["properties"].pop(key)
         metadata_schema["properties"]["Behavior"] = get_base_schema(tag="Behavior")
-        metadata_schema["properties"]["Behavior"].update(
-            required=["Videos"],
-            properties=dict(
-                Videos=dict(
-                    type="array",
-                    minItems=1,
-                    items=image_series_metadata_schema,
-                )
-            ),
+        metadata_schema["properties"]["Behavior"]["required"].append(self.metadata_key_name)
+        metadata_schema["properties"]["Behavior"]["properties"][self.metadata_key_name] = dict(
+            type="array",
+            minItems=1,
+            items=image_series_metadata_schema,
         )
         return metadata_schema
 
     def get_metadata(self):
         metadata = super().get_metadata()
-        behavior_metadata = dict(
-            Videos=[
+        behavior_metadata = {
+            self.metadata_key_name: [
                 dict(name=f"Video: {Path(file_path).stem}", description="Video recorded by camera.", unit="Frames")
                 for file_path in self.source_data["file_paths"]
             ]
-        )
+        }
         metadata["Behavior"] = behavior_metadata
 
         return metadata
@@ -307,9 +334,9 @@ class VideoInterface(BaseDataInterface):
         file_paths = self.source_data["file_paths"]
 
         # Be sure to copy metadata at this step to avoid mutating in-place
-        videos_metadata = deepcopy(metadata).get("Behavior", dict()).get("Videos", None)
+        videos_metadata = deepcopy(metadata).get("Behavior", dict()).get(self.metadata_key_name, None)
         if videos_metadata is None:
-            videos_metadata = deepcopy(self.get_metadata()["Behavior"]["Videos"])
+            videos_metadata = deepcopy(self.get_metadata()["Behavior"][self.metadata_key_name])
 
         assert len(videos_metadata) == self._number_of_files, (
             "Incomplete metadata "